Coordinate-long representations of geometries

R's graphics use coordinate-long representations of geometries (i.e., one row per coordinate). Additionally, these can be more easily manipulated than some other representations of geometry. Important: order matters when passing a geo_coord_*() to other functions, which assume that you use sequential feature, piece, and part values starting at 1. Helpers to construct this particular representation of geometry from a tibble are in the works.

geo_coord_linestring(xy, feature = 1L)

geo_coord_multilinestring(xy, feature = 1L, part = 1L)

geo_coord_point(xy, feature = seq_len(vec_size(xy)))

geo_coord_multipoint(xy, feature = seq_len(vec_size(xy)))

geo_coord_polygon(xy, feature = 1L, piece = 1L)

geo_coord_multipolygon(xy, feature = 1L, piece = 1L, part = 1L)

Arguments

xy
A geo_xy() (or something castable to geo_xy()) of coordinates

feature
An integer ID that whose unique values can be used to identify and/or separate features.

part
An ID used to separate parts within a feature of a multi* geometry

piece
An ID used to separate rings within polygons. The smallest piece value within a feature/part combination is assumed to be the outer ring.

Value

A vctr

Examples

geo_coord_point(geo_xy(30, 10))

#> <geo_coord_point [1 coords, 1 features]>
#> [1] <feat `1` (30 10)>

geo_coord_multipoint(geo_xy(1:4, 1:4), feature = c(1, 1, 2, 2))

#> <geo_coord_multipoint [4 coords, 2 features]>
#> [1] <feat `1.1` (1 1)> <feat `1.1` (2 2)> <feat `2.1` (3 3)> <feat `2.1` (4 4)>