Binary predicate operators

Binary predicate operators take two geometries and return TRUE or FALSE for each geometry pair.

geos_is_disjoint(x, y)

geos_touches(x, y)

geos_intersects(x, y)

geos_crosses(x, y)

geos_is_within(x, y)

geos_contains(x, y)

geos_overlaps(x, y)

geos_equals(x, y)

geos_covers(x, y)

geos_is_covered_by(x, y)

Arguments

x
A geometry-like object

y
A geometry-like object. x and y of length 1 is recycled to the length of the other (like tibble::tibble()).

Value

A logical vector

Details

geos_is_disjoint()

Returns TRUE if x and y do not have a point in common. Is the opposite of geos_intersects().

geos_touches()

Returns TRUE if x and y have a point in common, but their interiors do not intersect.

geos_intersects()

Returns TRUE if x and y have a point in common, regardless of whether or not their interiors intersect. Is the opposite of geos_is_disjoint().

geos_is_within()

Returns TRUE if all points in x are are also within y. geos_within(x, y) is identical to geos_contains(y, x).

geos_contains()

Returns TRUE if all points in y are are also within x. geos_contains(x, y) is identical to geos_within(y, x).

geos_overlaps()

Returns TRUE if x and y have some but not all points in common, and the intersection of the two is the same geometry type as the input. For example, two polygons can overlap, but a polygon and a line cannot.

geos_crosses()

Returns TRUE if x and y have some but not all points in common, and the intersection of the two is the not the same geometry type as the input. For example, a polygon and a line can cross, but two polygons cannot.

geos_equals()

Returns TRUE if x and y have all points in common and no points that intersect the exterior of the other.

geos_covers()

Returns TRUE if every point in y is a point of x. Is the opposite of geos_is_covered_by().

geos_is_covered_by()

Returns TRUE if every point in x is a point of y. Is the opposite of geos_covers().