Create geometries from data frames

Create geometries from data frames

unite_xy(data, col, x, y, remove = TRUE)

separate_xy(data, col, into = c("x", "y"), remove = TRUE)

Arguments

data
A tibble::tibble() or data frame

col
A column name for the geo_xy() object

x, y
Column specifications for x and y coordinate columns, respectively. Use dplyr::select() syntax.

remove
Use remove = FALSE to keep x and y as columns in the output.

into
Column names for the x and y coordinates, respectively.

Value

data, with a new column col

Examples

tbl <- tibble(x = 1, y = 2)
unite_xy(tbl, "xy", x, y, remove = FALSE)

#> # A tibble: 1 x 3
#>      xy     x     y
#>    <xy> <dbl> <dbl>
#> 1 (1 2)     1     2